* commit 'e6c9fb681f50d8d1b1d37de84f12d72b71e22ab7':
  Reverting the fields to long for backward compatibility

+++ business-logic/src/main/java/com/netflix/vms/transformer/SimpleTransformer.java
@@ -372,8 +372,8 @@ public class SimpleTransformer {
             while(iterator.hasNext()) {
                 PersonCharacterHollow personCharacterHollow = iterator.next();
                 MoviePersonCharacter moviePersonCharacter = new MoviePersonCharacter();
-                moviePersonCharacter.movieId = (int)movieId;
-                moviePersonCharacter.personId = (int)personCharacterHollow._getPersonId();
+                moviePersonCharacter.movieId = movieId;
+                moviePersonCharacter.personId = personCharacterHollow._getPersonId();
                 moviePersonCharacter.characterId = personCharacterHollow._getCharacterId();
                 personCharacters.add(moviePersonCharacter);
             }
